Update on Navbar items provider lookup app bar
