+++ src/App.css
@@ -4,6 +4,6 @@
   box-sizing: border-box;
   font-family: "Inter", sans-serif;
 }
-/* .App {
+.App {
   scroll-behavior: smooth;
-} */
+}
